Fix bug when search by pements, to avid diplicate lines.
when synch by cron, take the date from the last statement, not the
offset. Like the synch is only allowd a fixed days, if the certificate
is not update at the time, the next synch, may be the offset are not
enough, so ensure not left any bank line.

Update raise message givin more information.