Fix some bugs
Add the way to be able to control the refund invoices in a simple way.
Being able to associate them with the corresponding invoices and have them
pending payment again, to be able to follow them through the list of invoices,
claims (if the module is installed) and through the effects to be
paid/collected.

Update the origin form view.