On related_to field of Suggested Lines, remove part of the domain for the invoice and the account move line, to ensure that if an invoice is posted manually or an account move line is reconciled manually too, not produce an error. The suggested line is created throw the button, an this domains are taken in control there.
When set suggested lines to 'proposed' state, only set in the suggested
lines not in this state.

When post an origin ensure to delete the statement lines and suggested
lines with the same related_to.

When apply and invoice to multiple origins, ensure to remove the lines
and the suggested lines to avoid domain problems.

Task: #179421